+++ app/controllers/VilainController.java
@@ -22,8 +22,9 @@ import play.mvc.With;
 @With(AuthController.class)
 public class VilainController extends Controller {
 	
+	private static Utilisateur utilisateur = AuthController.connected();
+	
 	public static void index() {
-		Utilisateur utilisateur = AuthController.connected();
 		RolePermission permission = utilisateur.getPermission("VilainController", "read");
 		if (utilisateur.isAdmin || permission != null) {
 			 List<SurEtre> vilains = SurEtre.getSurEtreType(false);
@@ -38,7 +39,6 @@ public class VilainController extends Controller {
 	}
 	
 	public static void create() {
-		Utilisateur utilisateur = AuthController.connected();
 		if (utilisateur.can("VilainController", "create")) {
 	        String form = new Genform(new SurEtre(), "/vilain/add", "crudform").generate(validation.errorsMap(), flash);
 	        render("VilainController/form.html", form);
@@ -47,7 +47,6 @@ public class VilainController extends Controller {
     }
 	
 	public static void postCreate(@Valid SurEtre suretre) {
-		Utilisateur utilisateur = AuthController.connected();
 		if (utilisateur.can("VilainController", "create")) {
 			if(validation.hasErrors()) {
 	            params.flash();
@@ -69,45 +68,48 @@ public class VilainController extends Controller {
 		index();
     }
 
-	public static void update(long id) {
-		Utilisateur utilisateur = AuthController.connected();
-		SurEtre vilain = SurEtre.findById(id);
-		if (utilisateur.can("VilainController", "update")) {
-	        String form = new Genform(vilain, "/vilain/update/"+id, "crudform").generate(validation.errorsMap(), flash);
-	        render("VilainController/form.html", vilain, form);
+	public static void update(Long id) {
+		if(id != null) {
+			SurEtre vilain = SurEtre.findById(id);
+			if (utilisateur.can("VilainController", "update")) {
+		        String form = new Genform(vilain, "/vilain/update/"+id, "crudform").generate(validation.errorsMap(), flash);
+		        render("VilainController/form.html", vilain, form);
+			}
 		}
 		index();
 	}
 	
-	public static void postUpdate(long id) {
-		Utilisateur utilisateur = AuthController.connected();
-		SurEtre vilain = SurEtre.findById(id);
-		if (utilisateur.can("VilainController", "update")) {
-			if(validation.hasErrors()) {
-	            params.flash();
-	            validation.keep();
-	            update(id);
-	        }
-			Long[] avantages = params.get("suretre.avantages", Long[].class);
-			Long[] desavantages = params.get("suretre.desavantages", Long[].class);
-			if (avantages != null) {
-				vilain.avantages.addAll(Caracteristique.find("id in (?1)", Arrays.asList(avantages)).fetch());
+	public static void postUpdate(@Valid SurEtre suretre, Long id) {
+		if(id != null) {
+			suretre = SurEtre.findById(id);
+			if (utilisateur.can("SuperHeroController", "update")) {
+				suretre.edit(params.getRootParamNode(), "suretre");
+				suretre.civil = Civil.findById(params.get("suretre.civil", Long.class));
+				if(validation.hasErrors()) {
+		            params.flash();
+		            validation.keep();
+		            update(id);
+		        }
+				Long[] avantages = params.get("suretre.avantages", Long[].class);
+				Long[] desavantages = params.get("suretre.desavantages", Long[].class);
+				if (avantages != null) {
+					suretre.avantages.addAll(Caracteristique.find("id in (?1)", Arrays.asList(avantages)).fetch());
+				}
+				if (desavantages != null) {
+					suretre.desavantages.addAll(Caracteristique.find("id in (?1)", Arrays.asList(desavantages)).fetch());
+				}
+				suretre.save();
 			}
-			if (desavantages != null) {
-				vilain.desavantages.addAll(Caracteristique.find("id in (?1)", Arrays.asList(desavantages)).fetch());
-			}
-			vilain.edit(params.getRootParamNode(), "suretre");
-			vilain.civil = Civil.findById(params.get("suretre.civil", Long.class));
-			vilain.save();
 		}
 		index();
     }
 	
-	public static void delete(long id) {
-		Utilisateur utilisateur = AuthController.connected();
-		SurEtre vilain = SurEtre.findById(id);
-		if (utilisateur.can("VilainController", "delete")) {
-			vilain.delete();
+	public static void delete(Long id) {
+		if(id != null) {
+			SurEtre vilain = SurEtre.findById(id);
+			if (utilisateur.can("VilainController", "delete")) {
+				vilain.delete();
+			}
 		}
 		index();
 	}
